Started work on gcc compatability. Started to make Graphics components such as the uniforms and models more abstract using graphics provider and adapters (e.g. UniformAdapter). There are no more static Graphics instances. The TypedMap class has been renamed to the TypedSet, then true the TypedMap added. Extracted some code from the "meta/Meta.h" to separate headers. Improved sfinae checkers and "useif". Improved color support (RGB <-> HSV conversion, vector base). Added the "adapt" function to transmit parameters provided in variadic template functions. Improved handle ownership support.

+++ modules/freeimage/src/freeimage/FreeImageConverter.cpp
@@ -132,7 +132,7 @@ namespace Rapture
 		if(!FreeImage_SaveToMemory(internal_format, bmp, mem))
 			throw ImageConversionException("Can't save bitmap of format \"", type, "\" to memory");
 
-		unsigned long size;
+		DWORD size;
 
 		FreeImage_AcquireMemory(mem, &data, &size);
 		output->set(data, size);
